+++ src/Aardvark.Rendering.Vulkan/Runtime/Runtime.fs
@@ -128,8 +128,8 @@ type Runtime(device : Device) as this =
 
         device.CreateFramebuffer(unbox signature, views) :> IFramebuffer
 
-    member x.PrepareSurface (signature : IFramebufferSignature, surface : ISurface) =
-        device.CreateShaderProgram(unbox<RenderPass> signature, surface) :> IBackendSurface
+    member x.PrepareEffect (signature : IFramebufferSignature, effect : FShade.Effect) =
+        device.CreateShaderProgram(unbox<RenderPass> signature, effect, IndexedGeometryMode.TriangleList)
 
     member x.PrepareTexture (t : ITexture, [<Optional; DefaultParameterValue(false)>] export : bool) =
         ResourceValidation.Textures.validateForPrepare t
@@ -565,7 +565,7 @@ type Runtime(device : Device) as this =
         member x.CompileRender (signature, set) = x.CompileRender(signature, set)
         member x.CompileClear(signature, values) = x.CompileClear(signature, values)
 
-        member x.PrepareSurface(signature, s) = x.PrepareSurface(signature, s)
+        member x.PrepareEffect(signature, effect) = x.PrepareEffect(signature, effect)
         member x.PrepareRenderObject(fboSignature, rj) = x.PrepareRenderObject(fboSignature, rj)
         member x.PrepareTexture(t) = x.PrepareTexture(t)
         member x.PrepareBuffer(b, u, s) = x.PrepareBuffer(b, u, s)