+++ src/Aardvark.Rendering/Runtime/ResourceManager.fs
@@ -7,9 +7,6 @@ open FSharp.Data.Adaptive
 type IResourceManager =
 
     [<Obsolete("To be removed.")>]
-    abstract member CreateSurface : signature : IFramebufferSignature * surface : aval<ISurface> -> IResource<IBackendSurface>
-
-    [<Obsolete("To be removed.")>]
     abstract member CreateBuffer : buffer : aval<IBuffer> -> IResource<IBackendBuffer>
 
     [<Obsolete("To be removed.")>]