+++ src/Aardvark.Rendering/Surfaces/Surface.fs
@@ -4,17 +4,34 @@ open System
 open FShade
 open FSharp.Data.Adaptive
 
-[<AllowNullLiteral>]
-type ISurface = interface end
-
 type IBackendSurface =
-    inherit ISurface
     inherit IDisposable
     abstract member Handle : obj
+
+type DynamicSurface = EffectInputLayout * aval<Imperative.Module>
 
 [<RequireQualifiedAccess; ReferenceEquality>]
 type Surface =
     | FShadeSimple of Effect
-    | FShade of (EffectConfig -> EffectInputLayout * aval<Imperative.Module>)
-    | Backend of ISurface
+    | FShade of (EffectConfig -> DynamicSurface)
+    | Backend of IBackendSurface
     | None
+
+[<CompilationRepresentation(CompilationRepresentationFlags.ModuleSuffix)>]
+module Surface =
+
+    [<Sealed; AbstractClass>]
+    type Converter() =
+        static member inline ToSurface(surface: Surface) = surface
+        static member inline ToSurface(effect: Effect) = Surface.FShadeSimple effect
+        static member inline ToSurface(effects: #seq<Effect>) = Surface.FShadeSimple <| FShade.Effect.compose effects
+        static member inline ToSurface(create: EffectConfig -> DynamicSurface) = Surface.FShade create
+        static member inline ToSurface(surface: IBackendSurface) = Surface.Backend surface
+
+    let inline private toSurface (_ : ^Z) (data: ^T) =
+        ((^Z or ^T) : (static member ToSurface : ^T -> Surface)(data))
+
+    /// Creates a surface from the given data.
+    /// May be an effect, sequence of effects, backend surface, or dynamic surface function.
+    let inline create (data: ^T) =
+        toSurface Unchecked.defaultof<Converter> data